Unified jsonb tests
